requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yt-dlp>=2025.1.15
pydantic>=2.7.0
librosa>=0.10.2
torch>=2.3.0
torchaudio>=2.3.0
demucs>=4.0.0
openai-whisper>=20231117
basic-pitch>=1.6.0
